media: mfc: support block mode video playback
This supports block mode video playback and respects
the offset of each input stream.

+++ drivers/media/platform/exynos/mfc/mfc_data_struct.h
@@ -2040,8 +2040,8 @@ struct mfc_dec {
 	int crc_luma1;
 	int crc_chroma1;
 
-	unsigned long consumed;
-	unsigned long remained_size;
+	unsigned int consumed;
+	unsigned int remained_size;
 	dma_addr_t y_addr_for_pb;
 
 	int sei_parse;
